extract pure dao methods into classes
